// src/swagger-client/types.gen.ts
export type PageSize = {
    /**
     * in mm
     */
    height?: number;
    /**
     * in mm
     */
    width?: number;
};

export type RenderData = {
    /**
     * Optional html for footer. If empty, the footer html will be parsed from main html (<PdfFooter></PdfFooter>).
     */
    footerHtml?: string;
    /**
     * Optional html for header. If empty, the header html will be parsed from main html (<PdfHeader></PdfHeader>).
     */
    headerHtml?: string;
    html?: string;
    options?: RenderOptions;
};

export type RenderOptions = {
    excludeBuiltinStyles?: boolean;
    includeRenderLog?: boolean;
    landscape?: boolean;
    /**
     * margins in mm; fallback to default if null
     */
    margins?: RenderOptionsMargins;
    pageFormat?: 'A0' | 'A1' | 'A2' | 'A3' | 'A4' | 'A5' | 'A6' | 'Letter' | 'Legal';
    /**
     * page size in mm; overrides page format
     */
    pageSize?: PageSize;
};

export type RenderOptionsMargins = {
    /**
     * margin bottom in mm
     */
    bottom?: number;
    /**
     * margin left in mm
     */
    left?: number;
    /**
     * margin right in mm
     */
    right?: number;
    /**
     * margin top in mm
     */
    top?: number;
};

export type RenderTemplateData = {
    /**
     * Optional template for footer. If empty, the footer template will be parsed from main template (<PdfFooter></PdfFooter>).
     */
    footerHtmlTemplate?: string;
    /**
     * Optional template for header. If empty, the header template will be parsed from main template (<PdfHeader></PdfHeader>).
     */
    headerHtmlTemplate?: string;
    htmlTemplate?: string;
    /**
     * Model with your data matching to the templates
     */
    model?: {
        [key: string]: unknown;
    };
    options?: RenderOptions;
    templateEngine?: 'golang' | 'handlebars' | 'django';
};